Arcane logic
Currently the number of rows is hardcoded to my keyboard's [row
numbers](https://docs.qmk.fm/features/split_keyboard#layout-macro),
which means it will split for `0` to `5` as one side. Probably works
for most, but it is a limitation.

Note: My thumb cluster is one row per key, and two keys on each split.

The `set_arcane_cache` function runs on every key process, and
registers the keyevent of the keystroke if it is not `ARCANE`.
The logic of ignoring itself is inside the function.

The `process_arcane` functions takes the keyrecord as argument and
compares it with the `arcane_cache`. Based on the outcome it will
send repeated last keycode, or `Arcane` as input.