Arcane key on quantum level
The commit contains the first iteration of the Arcane key implemented
on quantum level.

The two modes are `Repeat/Cast` and `Leader`. Cast can be any adaptive custom
functionality launched from the other split side, whereas repeat happens
on the same split side.
`Repeat` and `Cast` reuse the repeat key logic.
The decision of `Repeat` or `Cast` is based on the number of the rows in
the matrix.
Note: That this is currently a hardcoded value.

The Leader mode is triggered by hitting the `ARCANE` outside of the
`LEADER_MODE_WAIT`.
When `arcane_leader` - so `ARCANE` is in leader mode, on key release it
starts a leader sequence utilizing the default leader logic.

Casts can be configured by the `process_arcane_cast_user(uint16_t keycode) {}`.